Add support for 'Authorization' header during token acquisition
See section [2.3.1](http://tools.ietf.org/html/rfc6749#section-2.3.1) of the latest version of the OAuth 2.0 specification. The library this is forked from only supports up to draft 10 of the specification and therefore does not support 'Authorization' headers as a means of passing client password to the server. Spring Security OAuth, however, expects it.